requirements.txt
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
defusedxml==0.7.1
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-Login==0.5.0
flask-mongoengine==1.0.0
Flask-OpenID==1.3.0
Flask-WTF==1.0.0
h11==0.12.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
lxml==4.6.4
MarkupSafe==2.0.1
mongoengine==0.23.1
outcome==1.1.0
pycparser==2.21
pymongo==3.12.1
pyOpenSSL==21.0.0
python-dotenv==0.19.2
python3-openid==3.2.0
requests==2.26.0
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
Werkzeug==2.0.2
wsproto==1.0.0
WTForms==3.0.0
